Move files.py under common.
