+++ Tools/c-analyzer/c_symbols/source.py
@@ -1,13 +1,41 @@
-from c_parser import files, declarations
+from c_analyzer_common import files
+from c_analyzer_common.info import UNKNOWN
+from c_parser import declarations
 
+
+# XXX need tests:
+# * find_symbol()
 
 def find_symbol(name, dirnames, *,
                 _perfilecache,
-                _get_local_symbols=(lambda *a, **k: _get_symbols(*a, **k)),
                 _iter_files=files.iter_files,
+                **kwargs
                 ):
     """Return (filename, funcname, vartype) for the matching Symbol."""
-    for filename in _iter_files(dirnames, ('.c', '.h')):
+    filenames = _iter_files(dirnames, ('.c', '.h'))
+    return _find_symbol(name, filenames, _perfilecache, **kwargs)
+
+
+def _get_symbols(filename, *,
+                 _iter_variables=declarations.iter_variables,
+                 ):
+    """Return the list of Symbols found in the given file."""
+    symbols = {}
+    for funcname, name, vartype in _iter_variables(filename):
+        if not funcname:
+            continue
+        try:
+            instances = symbols[name]
+        except KeyError:
+            instances = symbols[name] = []
+        instances.append((funcname, vartype))
+    return symbols
+
+
+def _find_symbol(name, filenames, _perfilecache, *,
+                _get_local_symbols=_get_symbols,
+                ):
+    for filename in filenames:
         try:
             symbols = _perfilecache[filename]
         except KeyError:
@@ -22,24 +50,9 @@ def find_symbol(name, dirnames, *,
         if not instances:
             symbols.pop(name)
         return filename, funcname, vartype
+    else:
+        return UNKONWN, UNKNOWN, UNKNOWN
 
 
-def _get_symbols(filename, *,
-                       _iter_variables=declarations.iter_variables,
-                       ):
-    """Return the list of Symbols found in the given file."""
-    symbols = {}
-    for funcname, name, vartype in _iter_variables(filename):
-        if not funcname:
-            continue
-        try:
-            instances = symbols[name]
-        except KeyError:
-            instances = symbols[name] = []
-        instances.append((funcname, vartype))
-    return symbols
-
-
-def iter_symbols(dirnames):
-    """Yield a Symbol for each symbol found in files under the directories."""
+def iter_symbols():
     raise NotImplementedError
